requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.9
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.9.1
attrs==25.3.0
boto3==1.38.23
botocore==1.38.23
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
cryptography==44.0.3
cssselect2==0.8.0
defusedxml==0.7.1
dj-database-url==2.3.0
Django==5.1.7
django-allauth==65.8.0
django-ckeditor==6.7.2
django-classy-tags==4.1.0
django-cms==4.1.4
django-entangled==0.5.4
django-filer==3.3.1
django-formtools==2.5.1
django-fsm==2.8.2
django-js-asset==3.1.2
django-parler==2.3
django-phonenumber-field==8.1.0
django-polymorphic==3.1.0
django-reversion==5.0.10
django-sekizai==4.1.0
django-storages==1.14.6
django-treebeard==4.7.1
djangocms-admin-style==3.3.1
djangocms-alias==2.0.2
djangocms-attributes-field==4.0.0
djangocms-frontend==1.3.5
djangocms-snippet==5.0.0
djangocms-text==0.7.1
djangocms-text-ckeditor==5.1.6
djangocms-versioning==2.2.1
easy-thumbnails==2.10
frozenlist==1.6.2
google-api-core==2.25.1
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
gunicorn==23.0.0
html5lib==1.1
httplib2==0.22.0
idna==3.10
jmespath==1.0.1
lxml==5.3.1
multidict==6.4.4
nh3==0.2.21
oauthlib==3.2.2
packaging==24.2
phonenumbers==9.0.6
pillow==11.1.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.31.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python3-openid==3.2.0
reportlab==4.3.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
s3transfer==0.13.0
setuptools==75.9.1
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
twilio==9.6.2
typing-extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
webencodings==0.5.1
whitenoise==6.9.0
yarl==1.20.0